requirements.txt
flask>=2.0.1,<3.0.0
werkzeug>=2.0.0,<3.0.0
gunicorn==20.1.0
google-cloud-videointelligence==2.11.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.35.0